Clear LEDs before setting spin animation.

+++ src/animation.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void AnimationController_initialise();
-void AnimationController_setAnimation(int animation, uint8_t* params, int length);
+void AnimationController_setAnimation(uint8_t animation, uint8_t* params, uint8_t length);
 
 #endif
